fixed warnings

+++ src/elf.cpp
@@ -961,7 +961,7 @@ u8 *elfReadAttribute(u8 *data, ELFAttr *attr)
     data += bytes;
     break;
   case DW_FORM_ref_addr:
-    attr->value = (elfDebugInfo->infodata + elfRead4Bytes(data)) - elfGetCompileUnitForData(data)->top;
+    attr->value = (u32)((elfDebugInfo->infodata + elfRead4Bytes(data)) - elfGetCompileUnitForData(data)->top);
     data += 4;
     break;
   case DW_FORM_ref4:
@@ -969,7 +969,7 @@ u8 *elfReadAttribute(u8 *data, ELFAttr *attr)
     data += 4;
     break;
   case DW_FORM_ref_udata:
-    attr->value = (elfDebugInfo->infodata + (elfGetCompileUnitForData(data)->top - elfDebugInfo->infodata) + elfReadLEB128(data, &bytes)) - elfCurrentUnit->top;
+    attr->value = (u32)((elfDebugInfo->infodata + (elfGetCompileUnitForData(data)->top - elfDebugInfo->infodata) + elfReadLEB128(data, &bytes)) - elfCurrentUnit->top);
     data += bytes;
     break;    
   case DW_FORM_indirect:
